requirements.txt
aiodns==3.1.1
aiohttp==3.9.3
aiosignal==1.3.1
alabaster==0.7.13
appdirs==1.4.4
arrow==1.2.3
astroid==2.15.6
asttokens==2.2.1
atomicwrites==1.4.1
attrs==23.1.0
audeer==1.20.1
audiofile==1.4.0
audioread==3.0.0
audmath==1.3.0
audtorch==0.6.4
autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
backtrader==1.9.78.123
baostock==0.8.8
beautifulsoup4==4.12.2
binance-connector==3.7.0
binance-futures-connector==4.0.0
binaryornot==0.4.4
black==23.7.0
bleach==6.0.0
blinker==1.6.2
blosc2==2.5.1
boltons @ file:///work/ci_py311/boltons_1677685195580/work
brotlipy==0.7.0
ccxt==4.2.84
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi @ file:///work/ci_py311/cffi_1676822533496/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.6
cloudpickle==2.2.1
cmake==3.27.0
comm==0.1.4
comtypes==1.2.0
conda @ file:///croot/conda_1689269889729/work
conda-content-trust @ file:///work/ci_py311/conda-content-trust_1676851327497/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1685032319139/work/src
conda-package-handling @ file:///croot/conda-package-handling_1685024767917/work
conda_package_streaming @ file:///croot/conda-package-streaming_1685019673878/work
contourpy==1.1.0
cookiecutter==2.3.0
cryptography @ file:///croot/cryptography_1686613057838/work
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
docstring-to-markdown==0.12
docutils==0.20.1
easyquotation==0.7.5
easytrader==0.23.0
easyutils==0.1.7
einops==0.6.1
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.13.3
flake8==6.0.0
Flask==2.3.2
fonttools==4.42.0
frozenlist==1.4.1
greenlet==2.0.2
h5py==3.9.0
idna @ file:///work/ci_py311/idna_1676822698822/work
imagesize==1.4.1
importlib-metadata==6.8.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.18.2
jeepney==0.8.0
jellyfish==1.0.0
Jinja2==3.1.2
joblib==1.3.2
jqdatasdk==1.9.0
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
keyring==24.2.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lazy_loader==0.3
libmambapy @ file:///croot/mamba-split_1685993156657/work/libmambapy
librosa==0.10.0.post2
lit==16.0.6
llvmlite==0.40.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.1
more-itertools==10.1.0
mpl-finance==0.10.1
mplfinance==0.12.10b0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
ndindex==1.8
nest-asyncio==1.5.7
networkx==3.1
numba==0.57.1
numexpr==2.9.0
numpy==1.24.4
numpydoc==1.5.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging @ file:///croot/packaging_1678965309396/work
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
ply==3.11
pooch==1.6.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycares==4.4.0
pycodestyle==2.10.0
pycosat @ file:///work/ci_py311/pycosat_1676838522308/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.20.0
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.15.1
pylint==2.17.5
pylint-venv==3.0.2
pyls-spyder==0.4.0
PyMySQL==1.1.0
pyOpenSSL @ file:///croot/pyopenssl_1678965284384/work
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyquery==2.0.0
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.8.2
python-lsp-black==1.3.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.4
python-slugify==8.0.1
pytoolconfig==1.2.5
pytz==2023.3
pywinauto==0.6.6
pyxdg==0.28
PyYAML==6.0.1
pyzmq==25.1.0
QDarkStyle==3.1
qstylizer==0.2.2
QtAwesome==1.2.3
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.1
requests==2.32.3
resampy==0.4.2
rich==13.5.2
rope==1.9.0
rpds-py==0.9.2
Rtree==1.0.1
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
scikit-learn==1.3.0
scipy==1.11.1
SecretStorage==3.3.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.4.5
spyder-kernels==2.4.4
SQLAlchemy==2.0.20
stack-data==0.6.2
sympy==1.12
tables==3.9.2
tabulate==0.9.0
text-unidecode==1.3
textdistance==4.5.0
threadpoolctl==3.2.0
three-merge==0.1.1
thriftpy2==0.4.16
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
toolz @ file:///work/ci_py311/toolz_1676827522705/work
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3 @ file:///croot/urllib3_1686163155763/work
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.3.6
whatthepatch==1.0.5
wrapt==1.15.0
wurlitzer==3.0.3
yapf==0.40.1
yarl==1.9.4
zipp==3.16.2
zstandard @ file:///work/ci_py311_2/zstandard_1679339489613/work
